+++ tests/test_functional.py
@@ -17,6 +17,7 @@ optimizers = [
     optim.Muon,
     optim.Lamb,
     optim.Kron,
+    optim.MARS,
     # optim.Shampoo,
 ]