+++ tests/test_neuralnet.py
@@ -3,7 +3,7 @@ from functools import partial
 import mlx.core as mx
 import mlx.nn as nn
 import pytest
-from mlx.optimizers import AdamW
+from mlx.optimizers import AdamW, clip_grad_norm
 
 import mlx_optimizers as optim
 
@@ -48,10 +48,14 @@ optimizers = [
         100,
     ),
     (optim.MADGRAD, {"learning_rate": 0.01}, 50),
-    (optim.ADOPT, {"learning_rate": 0.01}, 50),
+    (optim.ADOPT, {"learning_rate": 0.03}, 50),
     (optim.Lamb, {"learning_rate": 0.03}, 50),
     (optim.Shampoo, {"learning_rate": 0.03}, 50),
     (optim.Kron, {"learning_rate": 0.03}, 50),
+    (optim.MARS, {"learning_rate": 0.03, "mars_type": "mars-adamw"}, 50),
+    (optim.MARS, {"learning_rate": 0.03, "mars_type": "mars-lion"}, 50),
+    (optim.MARS, {"learning_rate": 0.03, "mars_type": "mars-shampoo"}, 50),
+    (optim.MARS, {"learning_rate": 0.03, "amsgrad": True}, 50),
 ]
 
 
@@ -74,6 +78,7 @@ def test_neuralnet(optimizer_config):
     def step(X, T):
         train_step_fn = nn.value_and_grad(model, eval_fn)
         loss, grads = train_step_fn(X, T)
+        grads, _ = clip_grad_norm(grads, 1)
         optimizer.update(model, grads)
         return loss
 
@@ -86,4 +91,5 @@ def test_neuralnet(optimizer_config):
             break
 
     acc = mx.sum(mx.argmax(model(X), axis=1) == T) / T.shape[0]  # type: ignore
-    assert losses[0] > 2 * losses[-1], f"loss={losses[-1]:.5f}, {acc=:.3f}"
+    assert losses[0] > 2 * losses[-1], f"Bad loss: loss={losses[-1]:.5f}, {acc=:.3f}"
+    assert acc > 0.85, f"Bad Acc: loss={losses[-1]:.5f}, {acc=:.3f}"